BUG1 : killed_process_rss = -1

+++ OperatingSystemPrj2/kernel/goldfish/mm/Yimin_oom.c
@@ -13,10 +13,11 @@ static unsigned long mm_current_list[e_num][2] = {{0}}; /* mm_current_list[][0]:
  * @uid: The uid of the app (of the entry u want to locate) 
  * @mm_current_list: The table which contains uid and its rss
  */
-int find_mm_current_list_index(unsigned long uid)
+static int find_mm_current_list_index(unsigned long uid)
 {
 	int i;
-	int availuable_index = -1;
+	int availuable_index;
+	availuable_index = -1;
 	for(i = 0; i < e_num; i++)
 	{
 		if(mm_current_list[i][0] == uid)
@@ -33,18 +34,22 @@ int find_mm_current_list_index(unsigned long uid)
  * __Yimin_kill - This function kill the process that has the highest RSS among all processes belonging to `uid`
  * @uid: The user id of the app who exceeds the memory limits set by syscall
  */
-void __Yimin_kill(uid_t uid,
-				  unsigned long usr_rss_before_killing, 
-				  unsigned long mm_max
-				 )
+static void __Yimin_kill(
+							uid_t uid,
+				  			unsigned long usr_rss_before_killing, 
+				  			unsigned long mm_max
+				 		)
 {
 	struct task_struct *iterator;
 	struct task_struct *p;
 	struct task_struct *killed_process = NULL;
-	unsigned long killed_process_rss = 0;
-	int kill_init_flag = 0;
+	unsigned long killed_process_rss;
 	pid_t killed_process_pid;
-
+	unsigned long pRSS_in_byte;
+	
+	killed_process_rss = -1;
+	pRSS_in_byte = 0;
+	
 	//* Step 1 - Find the process which has the highest RSS among all processes belonging to the user.
 	for_each_process(iterator)
 	{
@@ -61,7 +66,14 @@ void __Yimin_kill(uid_t uid,
 		task_unlock(p);
 	}
 
-	if(killed_process->pid == 1)
+	if (killed_process_rss == -1)
+	{
+		//! error handling
+		printk(KERN_ERR "killed_process_rss == -1 !\n");
+		return;
+	}
+
+	if (killed_process->pid == 1)
 	{
 		killed_process = NULL;
 		printk(KERN_ERR "> Refused to kill init (pid = 1) in Yimin's oom killer(which leads to kernel panic)...\n> Let's find the second biggest process...\n");
@@ -93,10 +105,10 @@ void __Yimin_kill(uid_t uid,
 	}
 
 	killed_process_pid = killed_process -> pid;
+
 	//* Step 2 - Kill the chosen process 
-	unsigned long pRSS_in_byte = 0;
 	pRSS_in_byte =  killed_process_rss * mm_page_size;
-	printk("Yimin's oom killer : uid = %d,\t uRSS = %luB,\t mm_max = %luB,\t pid = %d,\t pRSS = %luB", 
+	printk("Yimin's oom killer : uid = %lu,\t uRSS = %luB,\t mm_max = %luB,\t pid = %d,\t pRSS = %luB", 
 		    uid, usr_rss_before_killing, mm_max, killed_process_pid, pRSS_in_byte);
 	do_send_sig_info(SIGKILL, SEND_SIG_FORCED, killed_process, true);
 }
@@ -113,7 +125,7 @@ void __Yimin_kill(uid_t uid,
  * 			  If yes : kill the process that has the highest RSS among all processes belonging to the user.
  * 			  If no  : Do nothing and just return
  */
-void __Yimin_oom_killer()
+void __Yimin_oom_killer(void)
 {
 	struct task_struct *iterator;
 	struct task_struct *p;
@@ -122,6 +134,14 @@ void __Yimin_oom_killer()
 	int index;
 	int i;
 
+	extern struct Yimin_struct Yimin_mm_limits;
+	extern struct mutex Yimin_mutex;
+	int valid;
+	unsigned long mm_uid;
+	unsigned long mm_max;
+	unsigned long rss_in_byte;
+
+	rss_iter = 0;
 	//* Step 1 - Traverse all processes and collect the ones that are created by the same user.
 	for_each_process(iterator)
 	{
@@ -133,7 +153,7 @@ void __Yimin_oom_killer()
 		task_unlock(p);
 		if(rss_iter <= 0)
 			continue;
-		index = find_mm_current_list_index(uid_iter);
+		index = find_mm_current_list_index(uid_iter); 
 		//! error handling -> too many uid entries to handle
 		if(index == -1)
 		{
@@ -154,13 +174,6 @@ void __Yimin_oom_killer()
 	*/
 
 	//* Step 2 - Check memory limits set by syscall and kill some processes if needed
-	extern struct Yimin_struct Yimin_mm_limits;
-	extern struct mutex Yimin_mutex;
-	int valid;
-	unsigned long mm_uid;
-	unsigned long mm_max;
-	unsigned long rss_in_byte;
-
 	for(i = 0; i < e_num; i++)
 	{
 		mutex_lock(&Yimin_mutex); //Protect -> MMLimits (i.e. `Yimin_mm_limits` in my prj)
@@ -175,13 +188,13 @@ void __Yimin_oom_killer()
 		//* Valid entry in MMLimits
 		index = find_mm_current_list_index(mm_uid);
 		if(index == -1 || mm_current_list[index][1] == 0)
-			continue; //! entry is not currently running app
+			continue; //! Entry is not a currently running app
 
 		//* Valid entry in memory limits && entry is currently running
 		rss_in_byte = mm_current_list[index][1] * mm_page_size;
 		//printk(KERN_ERR "uid = %d,\t rss_in_byte = %luB\n", mm_uid, rss_in_byte);
 		if(mm_max < rss_in_byte){
-			printk(KERN_ERR "\nuid = %d, rss_in_byte = %luB  ----> Yimin_oom_killer triggered !\n", mm_uid, rss_in_byte);
+			printk(KERN_ERR "\nuid = %lu, rss_in_byte = %luB  ----> Yimin_oom_killer triggered !\n", mm_uid, rss_in_byte);
 			__Yimin_kill(mm_uid, rss_in_byte, mm_max);
 		}
 	}
